v1.7.8
- Correction de bugs avec les munitions multiples.
- Ajout de possibilités de désactiver les améliorations des armes à distance liées aux munitions (balles grappes, munitions explosives, ...) lors d'un jet
- Ajout d'une fenêtre de confirmation lors de la suppression d'un objet.

+++ module/sheets/actors/creature-sheet.mjs
@@ -4,7 +4,8 @@ import {
   getAEValue,
   listEffects,
   SortByName,
-  sum
+  sum,
+  confirmationDialog
 } from "../../helpers/common.mjs";
 
 import { KnightRollDialog } from "../../dialog/roll-dialog.mjs";
@@ -198,9 +199,11 @@ export class CreatureSheet extends ActorSheet {
       item.sheet.render(true);
     });
 
-    html.find('.item-delete').click(ev => {
+    html.find('.item-delete').click(async ev => {
       const header = $(ev.currentTarget).parents(".summary");
       const item = this.actor.items.get(header.data("item-id"));
+
+      if(!await confirmationDialog()) return;
 
       item.delete();
       header.slideUp(200, () => this.render(false));
