+++ module/sheets/actors/pnj-sheet.mjs
@@ -4,7 +4,8 @@ import {
   getAEValue,
   listEffects,
   SortByName,
-  sum
+  sum,
+  confirmationDialog
 } from "../../helpers/common.mjs";
 
 import { KnightRollDialog } from "../../dialog/roll-dialog.mjs";
@@ -204,9 +205,11 @@ export class PNJSheet extends ActorSheet {
       item.sheet.render(true);
     });
 
-    html.find('.item-delete').click(ev => {
+    html.find('.item-delete').click(async ev => {
       const header = $(ev.currentTarget).parents(".summary");
       const item = this.actor.items.get(header.data("item-id"));
+
+      if(!await confirmationDialog()) return;
 
       item.delete();
       header.slideUp(200, () => this.render(false));
